provisioner: add mutex to protect intercept config
Previously, a new collector or mediator could auth while
we were concurrently making modifications to the intercept
config via the update socket -- this may have had bad
ramifications for the set of intercepts etc that we would
have then sent to the collector/mediator.

Now, access to the intercept config is protected by a mutex
in both the update daemon thread and the main provisioner
epoll thread.