Make get_udp_payload even more general-purpose
It now takes two additional parameters which can be used
to return the source and dest ports of the UDP packet.

Updated the RADIUS plugin to use this new improved function,
instead of implementing its own version.

Updated other uses of the function to add the two new parameters,
which are set to NULL to tell the function that the callers don't
care about the port numbers.