GTP: remove IMSI from user ID
Intercepts are most likely to be specified using mobile numbers
and our user IDs are only useful if they can be directly matched
against whatever information the user can provide when starting
the intercept.

Phone number *should* be unique enough (?) and is more likely to
be readily available than IMSI, so will go with this approach
until informed otherwise.