GTP: properly match requests with responses
This is necessary for proper IRI construction, as some fields
are derived from request IEs and some are derived from the
response IEs.

At present, only Create Session and Delete Session pairs are
matched and considered for IRI generation, as these are the
best examples that I have so far.

No IRIs are actually generated as yet -- this code will come
in the near future.