Only attempt to flush pcap output files if we've written a packet
Files are not created (and therefore have a NULL file handle)
until libtrace is given a packet to write to them. Flushing a
NULL file handle can cause issues with the current packaged
versions of libtrace / libwandio, so we need to be extra
careful to not flush them unless we are sure there is something
to be flushed.