Allow multiple user identities from a IP sync packet
This work is only partially complete with this commit.

The main aim here is to allow users to have IP intercepts where
the target identity is found in the RADIUS Calling Station ID
rather than the Username.

In theory, this means we can have multiple identities come from
a single RADIUS packet (username + CSID) and it will be up to the
sync thread to decide which of these can be used for matching
against an active intercept.