Add "default RADIUS user" support
Operators that are using CSID as their primary user identity will
typically have a "default" username and password that they set
on all CPEs given to customers. These user names are meaningless
in terms of identity and we don't really want to be treating them
as a session key or potential target identity.

Added 'defaultradiususers' config option to intercept config, so
that operators can provide a list of RADIUS usernames that OpenLI
can safely ignore.

Note that operators don't *have* to set the 'defaultradiususers'
option, even if their customer devices are using a standard default
username in their RADIUS packets. Setting the option correctly
will help with performance, though, since OpenLI won't be trying
to track and update a session for the default username.