provisioner: re-enable reload of intercept config via SIGHUP
While the REST API is a better way to make on-the-fly config
changes, there are now users that have written tools that use the
SIGHUP feature to start and stop intercepts so it would be a bit
rude of me to remove that capability entirely.

Moved most of the config reloading code (for the provisioner)
into its own source file. This has meant that a number of
formerly static methods in provisioner.c are now available via
provisioner.h, as they're needed in both the reload-config code
path and the standard start-up code path.