Only replace intercept config if all YAML emitting was successful
Fixes #33.

When writing the updated intercept config after a REST API call,
we now write each bit of YAML to a temporary file, then copy
the contents of that file back over the top of the original
config file once all config has been successfully written to the
temporary file.

This fixes an issue where an error while writing the updated config
would result in the original config being lost and replaced with
whatever partial config we managed to write before hitting the
error.

Note we may still get an error when copying the temp file
content back into the original file, and that will still mess up
the original -- but that's a simpler operation with much less
likelihood of failure.