Added GET support to the REST API
For intercepts and agencies, you can either get the basic endpoint
to get a full list of all intercepts / agencies or get the
endpoint followed by a specific ID to get the details for a single
intercept / agency.

For coreservers and default RADIUS usernames, you can only get
the full list using the basic endpoint.