gtp: correctly match GTPv1 delete requests to their session
TEIDs for GTPv1 appear to be unidirectional, so our current
matching would miss any subsequent request messages for an
established session. End result: so IRI-Ends for GTPv1
sessions (and no removal of them from the session map
either).

TODO: update and test this to make sure it works for modify
requests as well.