Add collector config option to disable SDP O matching on VOIP calls
Background:

Section 5.3.1 of ETSI TS 102 232-5 makes a big deal about how
calls can be spread across multiple legs, where the Call ID
changes from leg to leg. It says that these legs should be
combined together under the same CIN and that one can do so
by looking for SIP messages where the O field inside the SDP
payload is identical.

This, unfortunately, is a rather optimistic view of how the O
field is used in practice. While in theory, the combination of
username, IP, version and session ID in the O field should be
unique for each call, there are many VOIP implementations that
are a bit lazy in how they assign session IDs (in particular),
resulting in multiple calls with the same O identifier.

This has resulted in situations where the O matching rules have
caused SIP packets for a completely different call to be lumped
in with a genuine intercept (which is very bad).

Since the SDP O matching is really only applicable in situations
where both sides of a SIP proxy are being fed into a single OpenLI
collector (unlikely), I've added a config option to allow
users to disable grouping calls together based on identical SDP O
data -- this means that operators using implementations that don't
do SDP O 100% correctly have the ability to tell OpenLI to not use
that information when looking for potential extra call legs.

Currently, the SDP O matching defaults to being on -- this may
change if bad SDP O fields become more common than not.