Add documentation and example for `sipignoresdpo` option
I've put in a pretty detailed explanation of why this option
exists and when a user might need to enable it. It's probably
a bit over the top and will go over the heads of most users,
but I think it will be helpful for anyone who needs to justify
enabling this option (e.g. if the LEA ends up complaining that
they have different CIN values for the IRIs generated from a
multi-leg call).