ETSILI IP addresses now make a copy of the IP address
Previously, we were just keeping the pointer to the address
that was passed in to the function. Problem is: for IRIs,
if the session object that the address came from is freed
before all relevant IRI messages involving that address are
encoded by the encoder workers, then we can end up encoding
bogus addresses.

For now, I free the copy of the address as soon as the
address is encoded -- this assumes that we are never going to
reuse that ETSILI IP address structure to do another encoding
later, so may need to revisit this if we do end up reusing it
for some reason.