radius: store some session properties for use in non-packet related IRIs
Specifically, this commit is intended to deal with #52.

We now store NAS identifier, NAS port, NAS IP, octets received
and octets transmitted against the radius_user_session_t for a
given session. This allows these values to be injected into IRIs
such as the startofInterceptionWithSessionActive IRI, where this
information would be useful to the IRI recipient (and may not
be available any other way).