First stages of adding digest-auth support to the provisioning REST API
The idea is to use an SQLCipher database to store login
credentials, which can then be queried by the update server to
check if the person sending a REST request is allowed to do so.

One problem is that libsqlcipher depends on libssl1.1 (on
stretch at least, probably also xenial and jessie). This conflicts
with our existing libcurl dependency (from wandio), which requires
libssl 1.0.2. Therefore, I've decided to only enable digest-auth
for hosts where the libssl is 1.1 or better; anyone running their
LI on something as old as stretch will just have to go without
this feature.