Fix a pile of compiler warnings
Most of these are just cases of unused variables left over from
previous versions of the code or meaningless signed vs unsigned
pointer complaints, but there were a handful of genuine bugs in
there too (e.g. '&' vs '&&' bugs). This shows the value of doing
this work to resolve the warnings -- to ensure genuine issues
don't get lost in a screed of compiler warnings, even if the
build succeeds.