collector: fix SSL write errors after a SSL_ERROR_WANT_WRITE
Turns out that we can only call SSL_write() again safely if we
pass in *exactly* the buffer and length that we used when we
got the WANT_WRITE error.

This solution does somewhat trap us in a blocking write state,
since we're going to keep trying until the waiting SSL write
succeeds. A better solution long-term would be to set the waiting
buffer aside and then start putting subsequent messages into a
new one. As long as we have a buffer set aside, try to write that
one first before moving on to the new one.