collector: make sure we fully free any pending encoding jobs on exit
liid and cinstr fields were not being properly freed, which was
showing up as a memory leak in valgrind (although this only
affected incomplete encoding jobs that were being drained from
the queues when the collector was exiting).

Also, when draining encoding jobs, make sure that we keep draining
until we get an error, rather than a simple EAGAIN. There may
still be more jobs on the way from the upstream threads that we
still need to drain and free before exiting -- the EAGAIN just
means that we're running ahead of the zmq thread(s).