More on-exit memory leak fixing
This commit updates the encoder workers to push a "null"
encoded result to each forwarding thread when they are exiting.

The forwarder threads can now use the receipt of this message
to know when there are going to be no more encoded results from
that encoder sitting in the queue.

Previously, we were relying on getting an EAGAIN when reading
from the queue to detect when there are no more results coming,
but this is not reliable (the encoders could just be a bit slower
than the forwarding thread). This meant that some encoded results
were not being properly freed on exit, leading to misleading
complaints about leaked memory in valgrind.