Add API key support to the REST authentication service
Instead of providing Digest-Auth, an API request can now also
include an "X-API-KEY" header. The value of this header must
match an "apikey" value in the authcreds table, but if it does
match then the request will be considered authenticated
automatically.

Obviously, it would be up to the users to secure any API keys
that they have, but it does make it easier to programatically
interface with an auth-protected provisioner without needing to
implement the full digest-auth process.