Add a couple of little bash scripts to help create a REST auth database
authsetup.sh simply creates an encrypted SQLite3 database with
a single table for storing auth credentials.

adduser.sh adds a single user to the aforementioned database table,
generating both a valid digest (based on the provided password)
and a random API key for that user.

The usage of these scripts will be documented on the OpenLI wiki.