Fix bug introduced by last bug fix :(
By returning early to avoid sending ETSI records over the socket
when we're meant to export to rabbitmq, we were also avoiding
the rather necessary code to complete a pending SSL handshake.

This would cause the mediator to never properly connect to a
collector, and therefore never actually receive any records when
SSL and RabbitMQ are enabled.