Halt components that are not using TLS when they should
As with the previous commit, this solves an issue where the
components would repeatedly reconnect without ever resolving
the root cause of the problem.

When a non-TLS client connects to an SSL-enabled provisioner,
the provisioner will now send an "SSL Required" message to the
client rather than immediately disconnecting the client. The
client should then receive that message and know that they should
immediately exit with an error message rather than reconnecting.

In the meantime, the provisioner will ignore any clients that
connect without using TLS and should eventually drop them as being
"idle".