Fix bug where reconnecting mediators would not get provisioner messages
The entry in the mediators map for the mediator had a pointer
to the old, disconnected client handle so that mediator would
always get skipped when we had a message to broadcast to all
mediators (such as a intercept announcement or new LIID mapping).

Now we make sure to replace the old handle with the entry from
the pendingclients map once the reconnection is confirmed.