Fix incorrect "CommunicationIdentifier" in HI1 Operations messages
The ETSI specs define CommunicationIdentifier in two different
places, with slightly different structures. Of course, we were
using the wrong one -- the one that we already use in the PS
Header, which is apparently different to the one defined in
the HI2 (!) specs that is meant to be used for HI1 Operations
messages.

Additional complications: the operator identifier in the
correct structure is limited to 5 characters, whereas the
operator identifiers that we have been using are meant to
be 16 characters max. So we now have to derive an appropriate
5 character identifier to put in the structure required by HI1.

The methods implemented are (in order of priority):
 * the user specifies the 5char identifier explicitly using
   the newly-added "altoperatorid" config option in the
   mediator config.
 * we use the first 5 characters from the existing operator
   ID provided in the mediator config.
 * we give up and set it to "?????".

Long story short: our encoding of the HI1 Operations messages
should be closer to being correct.