Add option to allow SIP From: URIs to be used for target identification
Normally we do not trust "From:" as it can be spoofed by the
sender and there is no in-protocol validation of it.

However, some VOIP networks may not be using the various auth
fields that we prefer in their SIP messages so it's nice for us
to at least provide an option for them to use "From:" if they
genuinely trust it.

The option defaults to "disabled", so users will have to explicitly
enable it, hopefully with the understanding of the risk of
spoofed identifiers.