export buffer: always attempt to send complete records
This fixes issues where a reconnecting client would be unable
to parse the data sent to it by the mediator because the
start of the first record had been sent prior to the disconnection
and was therefore lost.

Now we should only ever advance the buffer read pointer based on
a record boundary.

Tracking (and updating, in the case of a buffer realignment)
record boundaries will introduce a performance cost -- I am
still to evaluate the impact.

I've also increased the maximum send size per transmission from
16K to 1MB -- this should produce better performance without
realistically causing issues with blocking. The main consequence
is likely to be that we end up retransmitting more records that a
client has already seen in the case of a client reconnection, but
hopefully most downstream clients should be able to deal with
that themselves.