Save record boundaries less frequently for an export buffer
We don't need the precision of knowing the exact boundary of
every record -- the nearest record boundary every 32KB offers
more than enough precision for deciding where to cut off a
large send such that it ends on a record boundary.

This means we have a lot less saved offsets to re-calculate every
time we realign the buffer memory, and could even consider
increasing the gap between saved record boundary offsets
even further if need be.