mediator: periodically take a break from receiving from the collector
If the collector is sending us a ton of messages, we can get
stuck continuously reading these messages and never going back
to epoll to check for other events (particularly, the ability
to actually send some of those received messages on to the LEAs).

By forcing the receive method to end every 10 MB, we allow the
mediator the chance to maintain a relatively stable buffer of
received data.

The alternative is a potential death spiral of:
 * receive
 * receive
 * receive
 * buffer full -- resize buffer (this takes time, so more messages
   queue up to be received)
 * receive
 * receive
 * receive
 * buffer full again
 * etc.