Don't try to send data to handovers that failed their KA response
If a KA response timeout and a "transmit possible" event occur
for a handover on the same epoll wait loop, we could end up in a
situation where we disconnect and free the handover, then
immediately attempt to transmit data to the no-longer-existing
handover.

This could cause segfaults, or trigger an assertion due to a
bogus event type in the now-freed event object.