Add version numbers to encoding jobs
The version for an intercept is incremented each time that
a pre-encoded intercept field is modified, e.g. via a config
reload or a PUT request via the REST API.

The version is now part of the template identity key, so we
will create a fresh template for each intercept version rather
than incorrectly re-using a previously encoded template.

There is still an outstanding issue where templates are not
cleared when an individual intercept is removed -- this means
that if the same LIID is removed and then re-added, the version
number will be reset but the templates for any previous versions
will remain. This should be a pretty rare edge case, but ideally
we would handle this case properly.