Provisioner: treat digest hashes read from authdb as text
Our helper script inserts them as text but the
libmicrohttpd auth_check method needs the hash to be a binary
blob.

sqlite3_column_blob() wasn't doing the conversion from text to
binary as we had expected, so I've replaced this with code that
will do the conversion explicitly and ensure that the returned
digest is in the right binary format.