Fix SIP reassembly assertion bug
The reassembler could assert fail if the entire SIP header was
present, but some of the SDP payload was missing (i.e. in
another upcoming segment).

Instead of crashing, the reassembler should simply tell the caller
to wait for more payload, much like it does if the SIP header is
incomplete.