Better handling of SIP URIs that are just "sip:<IP addr>"
In this case, the username AND realm for the URI will be
assigned the value of <IP addr>, so it is possible for users
to intercept calls for that URI by simply setting the IP
address in question as their target username (and leave
realm blank).

This also means that seeing INVITES containing URIs in this
format won't cause OpenLI to complain about invalid SIP. It's
very unconventional SIP that probably won't interoperate very
well, but it might technically be valid so we should just try
to cope with it.