collector: fix bug where encoded raw IP was missing an LIID
At some point long ago, I modified all records exported to the
mediator to have an LIID string at the start to allow the
mediator to quickly look up the LIID without having to decode
ETSI content.

Apparently, I forgot to apply this change to exported "raw IP"
data (which is usually used to export raw GTP messages that
can be then written to pcap by the mediator for certain very
specific use cases).

This has been broken (and unnoticed) for quite a while, but I
finally ran into it myself when testing my mediator refactoring,
as raw IP was one of the features that I had to port over to the
refactored code.

Anyway, now this feature should finally work again -- and this is
also a good lesson in the value of some of these very specific
and obscure features ;)