Add payloadencryption as a common intercept field
Partially complete, still need to add support to the
provisioner json parser and throughout the collector
(i.e. for detecting changes in encryption method).

Refactored message construction and size calculations in netcomms
to add the "common" intercept fields in a single place, so we
don't have to change the code in multiple places every time we
add a new field. This probably introduces bugs, so make sure we
do plenty of testing with the various intercept types!