Fix issue where IPMM IRIs had incorrect IRI types
IRIs other than the first and last IRIs for a SIP session were
being tagged with the "Report" type, rather than the correct
"Continue" type.

Technically, "Report" is valid according to the spec but some
countries (e.g. DE) do not allow this option to be used (and
I never intended for us to abuse this loophole anyway!)