mediator postinst: set rabbitmq to listen on localhost only
Note: we only do this if we are the only users of rabbitmq
(i.e. it wasn't installed prior to our package being installed),
so we shouldn't break any existing rabbitmq setups.

But at the same time, a default RMQ install will be listening on
all interfaces and we don't need to be doing that for the use
case of our mediator. So let's keep it locked down if we can.