VOIP: fix bug where RTP stream info in 200 OK was ignored
If there was a 183 message that had no SDP information in it,
then we would not be able to recognise the subsequent 200 OK
that did have useful SDP info.

This is because we were deleting the saved CSEQ from the INVITE
after processing the 183, so we would never realise that the
200 OK was also in response to the INVITE (and therefore could
contain IPs and ports for a media stream).