imap: fix bug where IRIs were generated for a non-logged in session
If there was a session for which we had not seen the auth
exchange AND that session included a FETCH where the sender
was an intercept target, we would generate partial download
IRIs (but not CCs) for that FETCH command.

Since such IRIs without corresponding CCs is going to be
confusing for the authorities, we now do NOT generate IRIs for
these cases.

One could argue that if the sender was a target, then we should
generate both IRIs and CCs -- I think in practice that email
should be intercepted by a concurrent SMTP interception instead,
but we'll see what the agencies think if this ever becomes a
real issue.