Fix provisioner restart issues re: intercept config changes
If the intercept config was manually edited and the provisioner
was then restarted, some intercept config changes for existing
intercepts were not being passed on by the collector sync threads.

This was mainly caused by the code paths for handling a "modify"
intercept instruction and a "add already existing" intercept
instruction being different so it was easy to overlook one code
path when adding code to account for new intercept fields (e.g.
encryption options, or start/end times).

The two code paths have now been merged into one shared function
for each intercept type, so this should be less problematic in
the future.