Improve IMAP parsing for CRLF in replies
The main problem that this fixes is the appearance of unmatched
parentheses in message bodies that would have caused parsing
failures in the old code.

Changes:
 * use regex to match "{size}" patterns as these typically
   indicate blocks of non-IMAP protocol data (e.g. mail content)
   that should not be parsed. The "size" tells us how many bytes
   we need to skip to get back to IMAP protocol data, so we can
   move ahead to the next parseable byte.
 * simplify parsing flow into two states: all parentheses matched,
   or at least one open parenthesis. We only care about close
   parentheses and {size} patterns if we have at least one open
   parenthesis. We only care about \r\n if all the parentheses are
   closed.

Also included in this refactoring is a fix for a bug where the
collector could segfault if the reply was large and therefore the
final close parenthesis was not yet present in the buffer (because
it is coming in a subsequent packet).