REST API: better handling of empty string values for fields
Specifically, the encryption key can be an empty string for
unencrypted intercepts. We need to convert that into a NULL
value for that field in intercept_common_t so we don't try
to send 0 length strings to the collectors.

Also had to fix some of the intercept config modifying code to
now correctly handle cases where the value either was or is
changed to NULL.