Add a config option to control how compressed IMAP is turned into CCs
Options are "as-is", i.e. in compressed form, or "decompressed".

To support "as-is", the IMAP parser now retains the start and
end points for each observed packet within the compressed
stream, so we can send CCs with content that matches the original
packets from the wire.

Currently the default is "as-is", but this will also become a
configurable option.

Not entirely tested, but the default behaviour to send data
in its original compressed form seems to work well so far.