email ingestion: fix parsing failure due to packet segmentation
We were expecting libmicrohttpd to deliver a complete message to
the post iterator callback, but in fact it can deliver partial
messages so we need to do a little buffering of our own until
the remainder of the message arrives.

This bug was not obvious in my original testing, as I was using
virtual interfaces which meant that large messages were not being
segmented like they would be over a "real" interface.