collector: add config option to set default email domain
Email users that authenticate with just a username (and no
domain) will be assumed to have the email address
'<username>@<default domain>'.

If no default domain is set and we see a domainless email
username, we'll fallback to 'example.org' (which is not
particularly useful but will at least prevent things from
breaking).

Note, this option is specified in the collector config, not
the intercept config.