SIP: fix bug where a reverse INVITE would break RTP interception
SDP information in an INVITE was always being treated as describing
RTP for direction 0, and SDP information in 183, 200 replies was
always regarded as direction 1.

However, if an INVITE is sent in the reverse direction once the
call is connected (e.g. a codec update) then any SDP info in
either the INVITE or its 200 reply would overwrite the RTP for
the wrong direction.

This could lead to situations where an RTP stream that we had
determined to be from IP A, port X to IP B, port Y could then
be overwritten to be from IP A, port X to IP A, port X -- and
therefore interception of the actual RTP stream will cease.

Now we remember the IP address of the original "inviter" and
will be able to determine whether a subsequent INVITE is in the
reverse direction or not. So any changes to the RTP stream
should be correctly applied, and there should be no erroneous
updates either.