Add support for POP3 AUTH PLAIN commands
Moved some "common" code for handling PLAIN authentication into
email_worker.c so it can be used by both IMAP and POP3 parsers.

Note that any AUTH exchanges where the server sends challenges
is not going to be handled well right now, as this will
require a fair bit of work to refactor the POP3 parser to
account for commands that span more than a single client
request and server response.