Complete core functionality for SMS via SIP interception
Move the "trust_sip_from" configuration into the "shared"
configuration that threads have access to, so that the
corresponding behaviour will be correctly updated if the config
option is changed and then the config is reloaded.

Still TODO: handle the case where the intercept is IRI only,
and therefore we have to blank out the SMS message content
(and ONLY the content).