SMTP: disable TARGET_ID matching for ingested email content
TARGET_ID matching currently fails to produce correct results
in certain cases (e.g. automatic forwarding to an alias) -- we
will need to include additional information in the ingested
message format to help resolve these cases.

For now though, we get less errors without TARGET_ID matching
than we do with it so I've disabled it by default.

There is an undocumented config option to enable it -- in the
collector config, set "emailingest_usetargetid" to "on" to
re-enable TARGET_ID matching (not recommended). Once we resolve
the current problems with TARGET_ID matching, this option will
be documented and default to "on" instead.

Also fixed a bug where there was no working config option to
disable masking intercepted POP3 credentials.