IMAP/POP3: do not generate duplicate IRI events
Fetching mail where the target was the sender could create
duplicate "partial download" events -- one because sender ==
target and another because the target's mailbox was the
participant (i.e. recipient) for the mail session.

Previously, we never sent IRIs for recipients who were also the
sender of a mail but this is no longer the case, which has
introduced this bug.

Instead, all email events that only make sense in the context of
a mailbox now have a separate code path that ensures we only
create one IRI per mailbox per intercept that targets that mailbox.