Begin adding ability to include SIP packets in pcapdisk output
This is very incomplete and buggy at this stage, but the main
aim is to produce "rawip" encoding jobs for all SIP packets that
would normally be included as IRIs for an ETSI handover.

The complication comes when we have SIP messages that span
multiple packets or fragments, as the original packets were not
being retained. Now we try to keep copies of them alongside any
multi-fragment/segment SIP messages that we reassemble, and
return those back to the SIP parser when a complete message is
found.

Note to self: this is going to potentially introduce performance
issues (although the "fast" path used when a message fits in one
packet should result in no copying and be used almost all of the
time) -- we should only retain packet copies if there is an
active VOIP intercept that uses pcapdisk.