pcapdisk: only print file open messages once packet is written
libtrace doesn't actually create a pcap file until there is a
packet that has to be written into it (i.e. to avoid creating
a bunch of empty files when there is no traffic being captured).
Printing a log message every time trace_create() succeeds is
therefore a) spammy and b) misleading since the file may never
exist.

This change does mean that we have to store the filename in
the local pcap output structure, since we need it available
when we write our first packet into the file.