SMTP: add ability to detect automatic forwarding
The key result here is that, if the forwarded-from address is
a valid target, OpenLI will produce an email-send IRI for the
forwarding action with the forwarded-from address as the
sender.

This includes a configuration option for the collector to specify
what message header will contain the forwarded-from address
(e.g. "X-Sieve-Received-From"). If not specified, forwarding
detection will not occur.

There are still some caveats that need to be resolved:
 * only one forwarding message header type is supported, but it
   should really be a list of valid header fields
 * if the forwarded-to address is included in the target list for
   the intercept (i.e. it is an alias for the target), then we
   will also generate an email-receive IRI for the forwarding
   action -- this may not be the correct behaviour.
 * we make no distinction between forwarding detected on
   "outgoing" sessions (i.e. forwarding initiated by a user of
   the monitored mail service) and forwarding seen on an
   "incoming" session (i.e. the forward was initiated by an
   account on another mail provider).