Implement email forwarding headers as a set
Previous implementation only allowed the specification of one
email header that can be used to detect forwarding behaviour.

Now the user can provide a list of possible headers to check for
in the header section of the mail content.

As part of this, I've added a simple "string set" data structure
to util.c for managing sets of string terms (like the forwarding
headers).