gtp: avoid multiple calls to apply_gtp_fsm_logic per message
This fixes incorrect state changes that can occur if a
reordered request-response pair are processed multiple times
(i.e. once for each of MSISDN, IMSI, and IMEI).

The bad state change means that a) the original event is not
correctly reported for identities other than MSISDN and b)
any follow-up events can be missed for all identities.

But it turns out that we save the "action" from the most recent
call to apply_gtp_fsm_logic() in gparsed anyway, so we can just
use that.