Add support for port ranges for core server configuration
This allows users to tell OpenLI that a SIP service (for example)
can be running on any port number within the specified range, so
they can more easily account for dynamic port utilisation.

Single port core servers continue to be supported, and are simply
treated as a range where the upper and lower values are the same.

Other changes:
 * officially enable -Werror and -Wextra when compiling
 * refactor coreserver check code in collector.c to avoid
   duplicating much of the code already in coreserver.c