collector: do not send packets to sync threads if shutting down
This fixes an infinite loop that prevents a collector from halting
properly -- a processing thread wants to send a packet to a sync
thread, but the sync thread is waiting for the other threads to
halt and therefore not reading from the queue. If the queue is full,
then the processing thread blocks waiting for space to open up
but this will never happen because the sync thread has stopped
consuming.