collector: fix bug where session IP structs contained uninitialised bytes
This could potentially cause segfaults in the sync thread because
the full contents of the session IP struct are used as a hash key
to lookup active IPs -- if the session IP contained random bytes
on insertion then that will likely cause issues for later lookups
and deletions of active IPs.

Note: this bug only affects IP data sessions where the number of
assigned IPs exceeds 5, which I would expect to be a rare
situation.