collector SIP: don't insert invalid timer fds into timeouts map
If we fail to create a timer for the eventual timeout of a
completed call, the timer file descriptor returned will be -1.
This can happen if, for example, we run out of file descriptors.

Our timeouts map assumes timer fds are unique, so can become
corrupted if we end up inserting (and more importantly, deleting)
multiple entries using a key of -1.

So instead, if the timer fd is -1, we don't insert a timer into
the map -- instead, we just timeout the RTP streams for the call
immediately and avoid the potential memory corruption.

The recent fd leak fixes should prevent this from happening
anyway, but it doesn't hurt to also add an extra level of
protection just in case.