GTP worker threads now track TEIDs for GTP-U tunnels
Specifically, we are tracking the TEIDs for tunnels belonging
to intercept targets.

I've included the endpoint IPs for the tunnel to the identifier
for a given TEID, as a TEID is only unique for a particular link.
If a collector is provided traffic from multiple gateways, we will
hopefully avoid issues where the TEIDs clash.

Note to self: it might be better to use the IP address from inside
the F-TEID element but I would need to confirm how that IP maps
to addresses available in the GTP-U packets.