GTP: TEIDs are not bidirectionally unique
The only entity that matters for a TEID's uniqueness is the
destination IP. This means that the same TEID can be used for
both A->B and B->A paths, which would have broken our previous
approach.

This change should fix that problem, but needs to be properly
tested with traffic for the GTP-C and GTP-U channels from the
same session to confirm.