Resolve very rare RADIUS segfault
The core change here is to not automatically delete RADIUS users
who have no more active sessions.

The underlying explanation is a lot more complex...

The actual problem occurs when you have a combination of
conditions:
 * A RADIUS CSID that is used for more than one RADIUS username
   (or vice versa).
 * A little bit of packet loss on the RADIUS feed into the
   collector.
 * The Framed IP address for an affected shared CSID being
   re-assigned to another CSID without seeing the corresponding
   RADIUS session ending and start messages (either because of
   the packet loss, or because they were never mirrored in the
   first place).

What happens is that the IP reassignment triggers the sync
thread's "silent logoff" code branch. As part of that, all of
the identities previously attached to that IP address are
detached, and if that detached session was the last session for a
user then we would trigger the "delete user" code from the RADIUS
plugin.

Which make sense BUT the silent logoff removes ALL users as soon
as it is detected, and the silent logoff detection happens inside
a loop that is iterating over all users that have been reported as
belonging to that session.

So what if the user we just deleted is somehow in that list of
users that we are iterating over? Then we start playing around
with a user entity that we just freed and bad memory things will
happen.

But how does that user appear in the list if the IP has been
reassigned to another user?

That's where the non-unitary mapping of CSID to username comes
into play. The RADIUS plugin maintains a list of unanswered
requests, keyed by a combination of source port, request ID (8
bits) and request type. Requests are only deleted if they are
answered or if all the users involved in the request are deleted.
Normally, most requests are answered pretty promptly but if there
is packet loss then some responses may never be seen.

Now if the CSID has more sessions than the Username, any
unanswered requests associated with both the CSID AND the username
will persist, even if the username has no more sessions and was
therefore deleted.

If we then later on get an orphaned response (i.e. its original
request was also lost) that happens to match the key of one of
the old persisting requests then, guess what, we're going to dig
up the old request and start trying to update the session state
for the users that were originally involved with it.

Of course, if we happen to have deleted one of the users in the
interim (e.g. due to silent logoff, or just a regular session
ending) then bad things are going to happen again.

So ideally, we might want to think about having a way to expire
old requests (or at least ignore them when we're trying to match
a response) but this wouldn't help us with cases where we delete
a user early due to silent logoff. For now, the "safe" option is
not delete any users -- they're probably going to appear again
soon anyway -- with an eye towards instituting an expiry process
for seriously inactive users (and purging any requests still
associated with them, regardless of any other active users who
are also attached to those requests).